change to inline function
